+++ src/connection/BaseConnectionOptions.ts
@@ -25,7 +25,7 @@ export interface BaseConnectionOptions {
      * Accepts both entity classes and directories where from entities need to be loaded.
      * Directories support glob patterns.
      */
-    readonly entities?: (Function|string)[];
+    readonly entities?: ((Function|string|EntitySchema<any>))[];
 
     /**
      * Subscribers to be loaded for this connection.
@@ -33,13 +33,6 @@ export interface BaseConnectionOptions {
      * Directories support glob patterns.
      */
     readonly subscribers?: Function[]|string[];
-
-    /**
-     * Entity schemas to be loaded for this connection.
-     * Accepts both entity schema classes and directories where from entity schemas need to be loaded.
-     * Directories support glob patterns.
-     */
-    readonly entitySchemas?: EntitySchema[]|string[];
 
     /**
      * Migrations to be loaded for this connection.