+++ src/error/RepositoryNotFoundError.ts
@@ -1,12 +1,21 @@
+import {EntitySchema} from "../index";
+
 /**
  * Thrown when repository for the given class is not found.
  */
 export class RepositoryNotFoundError extends Error {
     name = "RepositoryNotFoundError";
 
-    constructor(connectionName: string, entityClass: Function|string) {
+    constructor(connectionName: string, entityClass: Function|EntitySchema<any>|string) {
         super();
-        const targetName = typeof entityClass === "function" && (<any> entityClass).name ? (<any> entityClass).name : entityClass;
+        let targetName: string;
+        if (entityClass instanceof EntitySchema) {
+            targetName = entityClass.options.name;
+        } else if (typeof entityClass === "function") {
+            targetName = entityClass.name;
+        } else {
+            targetName = entityClass;
+        }
         this.message = `No repository for "${targetName}" was found. Looks like this entity is not registered in ` +
             `current "${connectionName}" connection?`;
         Object.setPrototypeOf(this, RepositoryNotFoundError.prototype);