+++ src/error/RepositoryNotTreeError.ts
@@ -1,12 +1,21 @@
+import {EntitySchema} from "../index";
+
 /**
  * Thrown when repository for the given class is not found.
  */
 export class RepositoryNotTreeError extends Error {
     name = "RepositoryNotTreeError";
 
-    constructor(entityClass: Function|string) {
+    constructor(target: Function|EntitySchema<any>|string) {
         super();
-        const targetName = typeof entityClass === "function" && (<any> entityClass).name ? (<any> entityClass).name : entityClass;
+        let targetName: string;
+        if (target instanceof EntitySchema) {
+            targetName = target.options.name;
+        } else if (typeof target === "function") {
+            targetName = target.name;
+        } else {
+            targetName = target;
+        }
         this.message = `Repository of the "${targetName}" class is not a TreeRepository. Try to apply @Tree decorator on your entity.`;
         Object.setPrototypeOf(this, RepositoryNotTreeError.prototype);
         this.stack = new Error().stack;