+++ test/functional/entity-schema/indices/indices-test.ts
@@ -11,7 +11,7 @@ describe("indices > reading index from entity schema and updating database", () 
 
     let connections: Connection[];
     before(async () => connections = await createTestingConnections({
-        entitySchemas: [<any>PersonSchema],
+        entities: [<any>PersonSchema],
     }));
     beforeEach(() => reloadTestingDatabases(connections));
     after(() => closeTestingConnections(connections));